% Generated by roxygen2 (4.0.2): do not edit by hand
\name{deriv}
\alias{deriv}
\title{Derivatives}
\usage{
deriv(theta, items, resp, model = "GPCM", prior = NULL)
}
\arguments{
\item{theta}{Person parameter.}

\item{items}{Itembank, see \code{\link{genItembank}}.}

\item{resp}{Response pattern for the Itembank. Note that all items in the itembank need a corresponding response.}

\item{model}{IRT model, see \code{\link{MCAT}}. \strong{Defaults to GPCM}.}

\item{prior}{Covariance matrix if a prior should be included, \code{NULL} otherwise. \strong{Defaults to \code{NULL}}.
# Do we have to do something here for information? I.e. don't sum it?}
}
\value{
If \code{sum == TRUE} a list containing a matrix of first order derivatives,
and a list of matrices holding the second order derivatives per item.
otherwise a vector with the sum of first order derivatives and a matrix with the sum of second order derivatives. TODO: <- THIS
}
\description{
First and second order derivatives of the specified model, for the specified parameters.
}
\details{
Details
}

