% Generated by roxygen2 (4.0.2): do not edit by hand
\name{est}
\alias{est}
\title{Obtain an ability estimate given a response pattern and item parameters.}
\usage{
est(theta, items, resp, method = "BM", model = "GPCM", prior = NULL,
  threshold = 1e-04, max.iter = 10, debug = FALSE, ...)
}
\arguments{
\item{theta}{Numeric (vector), initial theta estimate. Defaults to a null vector of length equal to the number of dimensions.}

\item{items}{Itembank, see \code{\link{genItembank}}.}

\item{resp}{Response pattern, see \code{\link{ans}}.}

\item{method}{String, valid values are "ML", "EAP", and "MAP" or "BM". Defaults to "BM".}

\item{model}{String, valid values are "G3PLM", "GRM", "SM" or "GPCM". Defaults to "GPCM".}

\item{debug}{Logical. Should debug information be shown?}

\item{...}{Passed on to eapEst, ignored if method is not "EAP". See \code{\link{eapEst}} for details.}

\item{Sigma0}{Numeric (matrix), prior covariance. Only required when method != "ML".}
}
\value{
list with two entries; theta contains the current theta estimate, and SE the standard error of the estimate.
}
\description{
Calculates either the maximum likelihood (ML), estimated a-posteriori (EAP) or Bayes modal (BM or MAP) estimate.
}
\details{
Details
}

