% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/eapest.r
\name{eval.gauss}
\alias{eval.gauss}
\title{Evaluation of multivariate normal distributed integral}
\usage{
eval.gauss(FUN = function(x) 0, X = NULL, W = NULL, debug = FALSE, ...)
}
\arguments{
\item{FUN}{(Likelihood) function of the parameters to be estimated.
Defaults to \code{funtion(x) 1}, in which case only the built-in multivariate normal pdf is evaluated.}

\item{X}{Matrix of quadrature points, see \code{\link{init.gauss}}. Alternatively, the list of quadrature points and weights produced by \code{\link{init.gauss}}.}

\item{W}{Vector of weights, or \code{NULL} if provided in \code{X}.}

\item{...}{Additional arguments passed on to FUN.}

\item{prior}{Covariance matrix.}
}
\value{
A vector with the value evaluated integrals.
}
\description{
Evaluates a given function with a (built-in) multivariate normal prior distribution by Gauss-Hermite quadrature.
}
\details{
The evaluated function is assumed to have a multivariate normal distribution, with a given mean vector and covariance matrix.
The default identity function \code{function(x) 1} reduces to an integral over a multivariate normal distribution with mean vector \code{mu} and covariance matrix \code{Sigma}.
}
\examples{
quadPoints <- init.gauss(Q=3)
# expected value of 3-dimensional multivariate normal distribution: N(0,1).
# (Since mean is currently fixed at zero, this is always zero.)
(integral <- eval.gauss(Q=3,X=quadPoints))
round(integral)
}
\seealso{
\code{\link{init.gauss}} for creating quadrature points.
}

