% Generated by roxygen2 (4.0.2): do not edit by hand
\name{genItembank}
\alias{genItembank}
\title{Generate itembank.}
\usage{
genItembank(Q = 2, K = 50, model = "GPCM", covar = diag(Q),
  a = list(method = "normal", a = 0, b = 1), b = list(method = "normal", a =
  0, b = 1), c = list(a = 0, b = 0), M = list(min = 2, max = 4))
}
\arguments{
\item{Q}{integer; number of dimensions}

\item{K}{integer; number of items}

\item{model}{character; IRT model used to generate item parameters (can be either 'GRM', 'SM', or 'GPCM").}

\item{covar}{matrix; Covariance matrix for alpha (discrimination parameter), defaults to identity matrix of rank Q. See \bold{details}.}

\item{a}{list of options for sampling discrimination parameters. See \bold{details}.}

\item{b}{list of options for sampling difficulty parameters. See \bold{details}.}

\item{c}{list(min, max); named list of distribution parameters for the guessing parameter, c.
Guessing parameters are drawn from a continuous uniform distribution with minimum \code{a} and maximum \code{b}.
Default is list(min=0,max=0) for all \code{c} = 0.
Only used for G3PLM, ignored in all other models.}

\item{M}{list(min, max); named list of two integer values for the minimum and maximum values for a discrete uniform draw of number of categories per item. Ignored or G3PLM.}
}
\value{
Itembank. A list with class MCAT.items of parameters simulated under the given model.
    Additionally the itembank includes a list with options used to create it.
}
\description{
Generates a Q-dimensional itembank under the Graded Response Model (GRM), Sequential Model (SM), Generalized Partial Credit Model (GPCM) or Generalized Three Parameter Logistic Model (G3PLM).
}
\details{
Alpha parameters may be correlated, and are \code{> 0}.
This is accomplished by an initial draw from a multivariate normal distribution with the given covariance matrix.
We then take the exponent of these values, and for each item set the sum of exponents to 1.
\emph{This does mean that the observed covariance is not the same as the generating covariance matrix.}
}

