% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individuals.R
\docType{data}
\name{individuals}
\alias{individuals}
\title{Individual Mussel Samples, Upper Mississippi River, Lower Pool 10}
\format{A data frame with 1212 observations and 29 variables.
\describe{
   \item{OBJECTID}{numeric; The ArcGIS feature class assigned unique
                    identifier for each point in the dataset.}
   \item{SampleID}{}
   \item{Ename}{}
   \item{NumberLive}{}
}}
\source{
This dataset was downloaded from the \href
{https://rsgisias.crrel.usace.army.mil/intro/mussels_dev.pub.main}
{USACE Mussel Database}.
}
\usage{
individuals
}
\description{
Individual Mussel Samples, Upper Mississippi River, Lower Pool 10
}
\keyword{datasets}
