% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/information.R
\name{info}
\alias{info}
\title{Item information}
\usage{
info(theta, items, method = "FI", model = items$model,
  prior = items$prior)
}
\arguments{
\item{theta}{Numeric (vector) of latent abilitities.}

\item{items}{Itembank}

\item{method}{Information function to be used.}

\item{model}{IRT model to be used.}
}
\value{
information Numeric (vector) of information functions evaluated for the given item(s) and theta.
}
\description{
Computes item information for a given item and theta.
}
\details{
DETAILS
}

