% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/eapest.r
\name{init.gauss}
\alias{init.gauss}
\title{Initialize Gauss-Hermite quadrature points and weights.}
\usage{
init.gauss(ip = 10, prior)
}
\arguments{
\item{ip}{Integer, number of quadrature points, \bold{per dimension}. Note that the total number of quadrature points is \code{ip^Q}.}

\item{prior}{Covariance matrix of the prior distribution.}
}
\value{
X Matrix of dimensions \code{ip^Q} by \code{Q} of quadrature points.

W Vector of length \code{ip^Q} of weights.

ip Integer number of quadrature points per dimension used after sanity check.
}
\description{
Uses fastGHQuad to create unidemsional quadrature points.
These are subsequently expanded across the Q-dimensional grid, and shifted to account for the prior covariance matrix.
Weights are computed as the product of weights for each dimension at each quadrature point.
The mean vector is assumed to be zero, alternative prior means are not supported at this point.
}
\details{
Details
}
\seealso{
Package fastGHQuad, eval.gauss().
}

