% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MCAT.R
\name{initMCAT}
\alias{initMCAT}
\title{Initialise an MCAT object. This is the main test object, defining an MCAT. It includes an itembank, model, estimators and items selection criteria to be used,
as well as any constraints and starting/stopping criteria.}
\usage{
initMCAT(items = genItembank(model = model), model = "3PL", est = "BM",
  itemSelection = "PFI", start = list(type = "random", n = 5),
  stop = list(type = "length", n = 30), constraints = NULL)
}
\arguments{
\item{items}{MCAT_items object, see \code{\link{genItembank}} for a simulated itembank, and initItembank for using an existing itembank. Defaults to a simulated itembank.}

\item{model}{string, can be either "3PL", "GPCM", "GRM" or "SM". Defaults to "3PL".}
}
\description{
Initialise an MCAT object. This is the main test object, defining an MCAT. It includes an itembank, model, estimators and items selection criteria to be used,
as well as any constraints and starting/stopping criteria.
}

