% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MCAT.R
\name{initPerson}
\alias{initPerson}
\title{Initialise a person object for use in MCAT functions.}
\usage{
initPerson(items, theta = rep(0, items$Q))
}
\arguments{
\item{Items}{MCAT_items object, see \code{\link{genItembank}}. REQUIRED.}

\item{Theta}{numeric (vector) of ability levels, used to generate answers in simulations.}
}
\value{
Person person object, for use in \code{\link{MCAT:::MCAT}} and other functions.
}
\description{
Initialise a person object for use in MCAT functions.
}

