% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lampsilini.R
\docType{data}
\name{lampsilini}
\alias{lampsilini}
\title{Species of the Tribe Lampsilini}
\format{A data frame containing one variable.
\describe{
   \item{lampsilini}{A character vector containing the Genus and species of
                     mussels in the Tribe Lampsilini.}
}}
\source{
This dataset was taken from Heidi Dunn, Steve Zigler, and Teresa
Newton. 2016. Validation of a Mussel Community Assessment Tool for the
Upper Mississippi River System. 2014 MCA2. Rock Island, IL: U.S. Army
Corps of Engineers, Rock Island District.
}
\usage{
lampsilini
}
\description{
Species of the Tribe Lampsilini
}
\keyword{datasets}
