% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listed.R
\docType{data}
\name{listed}
\alias{listed}
\title{Listed Species}
\format{A data frame containing one variable.
\describe{
   \item{listed}{A character vector containing the Genus and species of
                 mussels that are Federal and State listed on the Upper
                 Mississippi.}
}}
\source{
This dataset was taken from Heidi Dunn, Steve Zigler, and Teresa
Newton. 2016. Validation of a Mussel Community Assessment Tool for the
Upper Mississippi River System. 2014 MCA2. Rock Island, IL: U.S. Army
Corps of Engineers, Rock Island District.
}
\usage{
listed
}
\description{
Listed Species
}
\keyword{datasets}
