\name{MCAT-package}
\alias{MCAT-package}
\alias{MCAT}
\docType{package}
\title{
What the package does (short line)
}
\description{
More about what it does (maybe more than one line)
}
\details{
\tabular{ll}{
Package: \tab MCAT\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-10-22\cr
License: \tab What license is it under?\cr
}
An overview of how to use the package, including the most important functions
}
\author{
Who wrote it

Maintainer: Who to complain to <yourfault@somewhere.net>
}
\references{
Literature or other references for background information
}
\keyword{ package }
\seealso{
\code{\link[<pkg>:<pkg>-package]{<pkg>}}
}
\examples{
print("BOO!")
}
