% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{mcat}
\alias{mcat}
\alias{mcat-package}
\title{A Package for Calculating Mussel Cummunity Assessment Tool Metrics}
\description{
This package implements the Mussel Community Assessment Tool (MCAT)
        metrics in Heidi Dunn, Steve Zigler, and Teresa Newton. 2016. 
        “Validation of a Mussel Community Assessment Tool for the Upper 
        Mississippi River System.” 2014 MCA2. Rock Island, IL: U.S. Army Corps 
        of Engineers, Rock Island District.
}
\details{
This package implements the Mussel Community Assessment Tool (MCAT)
    metrics in Heidi Dunn, Steve Zigler, and Teresa Newton. 2016.
    “Validation of a Mussel Community Assessment Tool for the Upper
    Mississippi River System.” 2014 MCA2. Rock Island, IL: U.S. Army Corps
    of Engineers, Rock Island District.
}
\keyword{internal}
