% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcat_site_score.R
\name{mcat_site_score}
\alias{mcat_site_score}
\title{MCAT Site Score}
\usage{
mcat_site_score(individuals, threshold = default_threshold)
}
\arguments{
\item{individuals}{data frame; a data frame of individual mussel records.}

\item{threshold}{a Thresold object; An object that holds the thresholds
for each MCAT metric above which a site is considered
to have met the MCAT metric threshold for consideration
as a healthy mussel community for that metric.}
}
\value{
A data frame of sampled sites with the calculated suite of MCAT
        metrics and a total site score.
}
\description{
Calculates the MCAT site score for the input individual mussel
data frame.
}
\examples{
# Create the individual mussel data frame
individuals <- mcat::individuals

# Calculate percent fresh dead for the individuals data frame
mss <- mcat_site_score(individuals)

}
