% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MCAT_helpers.R
\name{next_item}
\alias{next_item}
\title{Select the next item, given a test and person}
\usage{
next_item(MCAT, person, debug = FALSE)
}
\arguments{
\item{MCAT}{test object, see \code{\link{initMCAT}}.}

\item{person}{MCAT_person object, see \code{\link{initPerson}}}

\item{debug}{boolean, print debug info?}
}
\description{
Returns the next item index, based on item selection criteria, constraints and other applicable criteria.
}

