% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent_fresh_dead.R
\name{percent_fresh_dead}
\alias{percent_fresh_dead}
\title{Percent Fresh Dead}
\usage{
percent_fresh_dead(individuals)
}
\arguments{
\item{individuals}{data frame; a data frame of individual mussel records.}
}
\value{
A data frame of sampled sites with the calculated percent fresh dead
        MCAT metric
}
\description{
Calculates the percent fresh dead MCAT metric for the input
individual mussel data frame.
}
\examples{
# Create the individual mussel data frame
individuals <- mcat::individuals

# Calculate percent fresh dead for the individuals data frame
pj <- percent_juveniles(individuals)

}
