% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent_lampsilini.R
\name{percent_lampsilini}
\alias{percent_lampsilini}
\title{Percent Tribe Lampsilini}
\usage{
percent_lampsilini(individuals)
}
\arguments{
\item{individuals}{data frame; a data frame of individual mussel records.}
}
\value{
A data frame of sampled sites with the calculated percent tribe
        lampsilini MCAT metric
}
\description{
Calculates the percent tribe Lampsilini MCAT metric for the
input individual mussel data frame.
}
\examples{
# Create the individual mussel data frame
individuals <- mcat::individuals

# Calculate percent tribe lampsilini for the individuals data frame
plm <- percent_lampsilini(individuals)

}
