% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent_over_15yrs.R
\name{percent_over_15yrs}
\alias{percent_over_15yrs}
\title{Percent Over 15 Years}
\usage{
percent_over_15yrs(individuals)
}
\arguments{
\item{individuals}{data frame; a data frame of individual mussel records.}
}
\value{
A data frame of sampled sites with the calculated percent over 15 years
        MCAT metric
}
\description{
Calculates the percent over 15 years MCAT metric for the input
individual mussel data frame.
}
\examples{
# Create the individual mussel data frame
individuals <- mcat::individuals

# Calculate percent over 15 years for the individuals data frame
po <- percent_over_15yrs(individuals)

}
