% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent_tolerant.R
\name{percent_tolerant}
\alias{percent_tolerant}
\title{Percent Tolerant}
\usage{
percent_tolerant(individuals)
}
\arguments{
\item{individuals}{data frame; a data frame of individual mussel records.}
}
\value{
A data frame of sampled sites with the calculated percent tolerant
        MCAT metric
}
\description{
Calculates the percent tolerant MCAT metric for the input
individual mussel data frame.
}
\examples{
# Create the individual mussel data frame
individuals <- mcat::individuals

# Calculate percent tolerant for the individuals data frame
pt <- percent_tolerant(individuals)

}
