% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/genitems.r
\name{plot.MCAT_itembank}
\alias{plot.MCAT_itembank}
\title{Plot itembank with some useful detail.}
\usage{
\method{plot}{MCAT_itembank}(itembank)
}
\description{
Plot itembank with some useful detail.
}

