% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/genitems.r
\name{print.MCAT.items}
\alias{print.MCAT.items}
\title{Print itembank with some useful detail.}
\usage{
\method{print}{MCAT.items}(itembank)
}
\description{
Print itembank with some useful detail.
}

