% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/genitems.r
\name{print.MCAT_itembank}
\alias{print.MCAT_itembank}
\title{Print itembank with some useful detail.}
\usage{
\method{print}{MCAT_itembank}(itembank)
}
\description{
Print itembank with some useful detail.
}

