% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/prob.R
\name{prob}
\alias{prob}
\title{Probability of answer given item and theta.}
\usage{
prob(theta = 0, items, model = items$model)
}
\arguments{
\item{theta}{Numeric, vector of length Q ability parameters.}

\item{items}{Itembank, or a subset of items from an itembank. See also \code{\link{genItembank}}}

\item{model}{Character, IRT response model. Can be any of "G3PL", "GRM", "SM" or "GPCM".}
}
\value{
List with two entries, a vector or matrix \code{p} of probabilites per item/category,
and a vector \code{p0} of probabilities defined as \code{p0 = q = 1 - sum(P_ij)}
}
\description{
Returns a vector (for G3PLM) or a matrix (GRM,SM or GPCM) of probabilities for one or more items,
given a specified ability (vector).
}
\details{
Detailed description, response models.
}

