% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MCAT.R
\name{SimulateMCAT}
\alias{SimulateMCAT}
\title{Simulate an MCAT for a single person.}
\usage{
SimulateMCAT(person, items, st = 5, le = 50, criteria = NULL,
  debug = FALSE)
}
\arguments{
\item{person}{List of person parameters and related information. See \code{\link{MCAT::person}}.}

\item{criteria}{List of test criteria as defined by \code{\link{MCAT::MCAT}}.}

\item{itembank}{List of item parameters and related information. See \code{\link{MCAT::genItembank}}.}
}
\value{
person
}
\description{
Wrapper to simulate an MCAT with the given criteria for a given person and itembank.
}
\details{
DETAILS (note that only PFI/FI objective functions are implemented, and only for 3PL)
}

