% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_evenness.R
\name{species_evenness}
\alias{species_evenness}
\title{Species Evenness}
\usage{
species_evenness(individuals)
}
\arguments{
\item{individuals}{data frame; a data frame of individual mussel records.}
}
\value{
A data frame of sampled sites with the calculated species evenness
        MCAT metric
}
\description{
Calculates the species evenness MCAT metric for the input
individual mussel data frame.
}
\examples{
# Create the individual mussel data frame
individuals <- mcat::individuals

# Calculate species evenness for the individuals data frame
se <- species_evenness(individuals)

}
