% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_richness.R
\name{species_richness}
\alias{species_richness}
\title{Species Richness}
\usage{
species_richness(individuals)
}
\arguments{
\item{individuals}{data frame; a data frame of individual mussel records.}
}
\value{
A data frame of sampled sites with the calculated species richness
        MCAT metric
}
\description{
Calculates the species richness MCAT metric for the input
individual mussel data frame.
}
\examples{
# Create the individual mussel data frame
individuals <- mcat::individuals

# Calculate species richness for the individuals data frame
sr <- species_richness(individuals)

}
