% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/genitems.r
\name{subset.MCAT_itembank}
\alias{subset.MCAT_itembank}
\title{Create a subset of the itembank containing the given items.}
\usage{
\method{subset}{MCAT_itembank}(itembank, ss)
}
\arguments{
\item{itembank}{MCAT_items, list with class MCAT_items, see genItembank}

\item{ss,}{vector of integers, indeces of item(s) to be included in the subset.}
}
\value{
MCAT_itembank, subset containing item(s) ss and all the other bits and pieces that make an itembank.
}
\description{
Create a subset of the itembank containing the given items.
}

