% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Threshold.R
\docType{class}
\name{Threshold-class}
\alias{Threshold-class}
\title{Threshold}
\description{
The Threshold object is designed to hold the cut-off values for
each MCAT metric above which a site is considered to have met the MCAT
metric threshold for consideration as a healthy mussel community for that
metric.
}
\section{Slots}{

\describe{
\item{\code{species_richness}}{}
}}

