% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tribe_evenness.R
\name{tribe_evenness}
\alias{tribe_evenness}
\title{Tribe Evenness}
\usage{
tribe_evenness(individuals)
}
\arguments{
\item{individuals}{data frame; a data frame of individual mussel records.}
}
\value{
A data frame of sampled sites with the calculated tribe evenness
        MCAT metric
}
\description{
Calculates the tribe evenness MCAT metric for the input
individual mussel data frame.
}
\examples{
# Create the individual mussel data frame
individuals <- mcat::individuals

# Calculate tribe evenness for the individuals data frame
te <- tribe_evenness(individuals)

}
