% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MCAT_helpers.R
\name{update.MCAT_person}
\alias{update.MCAT_person}
\title{Update person object after a response. Update administrative variables, and calculate new estimates and variance.}
\usage{
\method{update}{MCAT_person}(person, MCAT, item, answer)
}
\arguments{
\item{person}{MCAT_person object, see \code{\link{initPerson}}}

\item{MCAT}{MCAT test object, see \code{\link{initMCAT}}.}

\item{item}{integer, index of item responded to.}

\item{answer}{integer, actual response.}

\item{...}{parameters passed onto the estimator. threshold and max.iter influence BM and ML Newton-Rhapson estimation routine, ip sets the number of quadrature points per dimension in EAP estimation.}
}
\value{
MCAT_person updated person object.
}
\description{
Update person object after a response. Update administrative variables, and calculate new estimates and variance.
}

