% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MCAT_helpers.R
\name{update_response}
\alias{update_response}
\title{Update person object after a response. Update administrative variables, and calculate new estimates and variance.}
\usage{
update_response(MCAT, person, item, answer)
}
\arguments{
\item{MCAT}{test object, see \code{\link{initMCAT}}.}

\item{person}{MCAT_person object, see \code{\link{initPerson}}}

\item{item}{Integer, index of item responded to.}

\item{answer}{Integer, actual response.}
}
\value{
person Updated MCAT_person object.
}
\description{
Update person object after a response. Update administrative variables, and calculate new estimates and variance.
}

