% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ans}
\alias{ans}
\title{Simulate a response (pattern) with a given theta and item(s)}
\usage{
ans(theta, items, model = "GPCM")
}
\arguments{
\item{theta}{Numeric (vector). Vector of length Q, where each element is the theta for the corresponding dimension}

\item{items}{Itembank, or a subset of an itembank. See \code{\link{genItembank}} and \code{\link{subset.MCAT.items}}}
}
\value{
integer (vector) of simulated response(s).
}
\description{
Simulate a response (pattern) with a given theta and item(s)
}

