% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/eapest.r
\name{eapEst}
\alias{eapEst}
\title{EAP estimate of person parameters.}
\usage{
eapEst(items, resp, model = items$model, prior = items$covar, ip = 4,
  debug = FALSE, ...)
}
\arguments{
\item{items}{Itembank, or a subset of an Itembank. See \code{\link{genItembank}}.}

\item{resp}{Response pattern, see \code{\link{ans}}.}

\item{model}{String, valid values are "3PL", "GRM", "SM" or "GPCM". Defaults to "GPCM".}

\item{prior}{Numeric (matrix), prior covariance.}

\item{debug}{Logical. Should debug information be shown?}

\item{theta}{Numeric (vector), initial theta estimate. Defaults to a null vector of length equal to the number of latent dimensions.}

\item{quad}{Integer, number of quadrature points \bold{per dimension}. Note that the total number of quadrature points is \code{quad^Q}, where Q is the number of latent dimensions.}
}
\value{
estimate A vector of latent ability estimates.
}
\description{
Estimated-a-posteriori estimate of latent ability. Utilizes Gauss-Hermite quadrature to do the required numerical integration.
Can provide better estimates than ML estimation, particularly early on in a test.
}
\details{
Details
}
\seealso{
eapSE.
}

