% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individuals.R
\docType{data}
\name{individuals}
\alias{individuals}
\title{Individual Mussel Samples, Upper Mississippi River, Lower Pool 10}
\format{A data frame with 1212 observations and 29 variables.
\describe{
   \item{SampleID}{}
   \item{Ename}{}
   \item{Previously_Marked}{}
   \item{Marked_mussel}{}
   \item{Status}{}
   \item{Sex}{}
   \item{Gravidity}{}
   \item{Age}{}
   \item{Length}{}
   \item{Height}{}
   \item{Voucher_Location}{}
   \item{Voucher_Type}{}
   \item{ZebrasAttached}{}
   \item{Comments}{}
   \item{Northing}{}
   \item{Easting}{}
   \item{Method}{}
   \item{NumberLive}{}
   \item{NUM5YEAR}{}
   \item{NumZebraAttached}{}
   \item{PerZebraCoverage}{}
}}
\source{
This dataset was downloaded from the \href
{https://rsgisias.crrel.usace.army.mil/intro/mussels_dev.pub.main}
{USACE Mussel Database}.
}
\usage{
individuals
}
\description{
Individual Mussel Samples, Upper Mississippi River, Lower Pool 10
}
\keyword{datasets}
