% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/MCAT.R
\name{MCAT}
\alias{MCAT}
\title{Simulate an MCAT for a single person.}
\usage{
MCAT(person, items, criteria = NULL, debug = FALSE)
}
\arguments{
\item{person}{List of person parameters and related information. See \code{\link{MCAT::person}}.}

\item{criteria}{List of test criteria as defined by \code{\link{MCAT::MCAT}}.}

\item{itembank}{List of item parameters and related information. See \code{\link{MCAT::genItembank}}.}
}
\value{
person
}
\description{
Wrapper to simulate an MCAT with the given criteria for a given person and itembank.
}
\details{
DETAILS (note that only PFI/FI are implemented, and only for 3PL)
}

