% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Threshold.R
\docType{class}
\name{Threshold-class}
\alias{Threshold-class}
\title{Threshold}
\description{
The Threshold object is designed to hold the cut-off values for
each MCAT metric above which a site is considered to have met the MCAT
metric threshold for consideration as a healthy mussel community for that
metric.
}
\section{Slots}{

\describe{
\item{\code{percent_listed}}{numeric; Percent of mussels listed}

\item{\code{percent_tolerant}}{numeric; Percent of mussels tolerant}

\item{\code{percent_lampsilini}}{numeric; Percent of mussels lampsilini}

\item{\code{percent_juveniles}}{numeric; Percent of mussels juveniles}

\item{\code{percent_over_15yrs}}{numeric; Percent of mussels over 15 years old}

\item{\code{percent_fresh_dead}}{numeric; Percent of mussels freshly dead}

\item{\code{abundance}}{numeric; Abundance of mussels}

\item{\code{species_evenness}}{numeric; Species evenness of mussels}

\item{\code{tribe_evenness}}{numeric; Tribe evenness of mussels}

\item{\code{species_richness}}{numeric; Species richness of mussels}
}}

